client/requirements.txt
aiohttp
python-dotenv
elasticsearch<9.0
